fixed build on mingw. Lua still doesn't work on Linux. have to fix Lua build itself

+++ src/modules/modules.h
@@ -11,16 +11,16 @@
 #pragma once
 
 extern unsigned char module_bus_lua[];
-extern size_t module_bus_lua_len;
+extern unsigned long long module_bus_lua_len;
 
 extern unsigned char module_custom_lua[];
-extern size_t module_custom_lua_len;
+extern unsigned long long module_custom_lua_len;
 
 extern unsigned char module_events_lua[];
-extern size_t module_events_lua_len;
+extern unsigned long long module_events_lua_len;
 
 extern unsigned char module_pin_lua[];
-extern size_t module_pin_lua_len;
+extern unsigned long long module_pin_lua_len;
 
 extern unsigned char device_lua[];
-extern size_t device_lua_len;
+extern unsigned long long device_lua_len;
